Move app module to the app folder
