+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Organize file structure.
+
 ## [0.1.0] - 2019-08-01
 
 ### Added
 
-- Initial version
+- Initial version.
